+++ project_config.yml
@@ -9,7 +9,7 @@ acc:
 dev:
   catalog_name: mlops_dev
   schema_name: acikgozm
-  pipeline_id: 0498db79-f551-49bb-ad4f-finac9d91fc
+  pipeline_id: e836641e-403e-4b28-8622-c6d1ed937c59
 
 #default values for notebook implementations
 # catalog_name: mlops_dev
