+++ scripts/01_data_load_preprocess_script.py
@@ -7,7 +7,8 @@ from dotenv import load_dotenv
 from loguru import logger
 
 from hotel_reservations.config import Config
-from hotel_reservations.data_ingestion import DataFabricator, DataLoader
+from hotel_reservations.data_ingestion import DataLoader
+from hotel_reservations.synthetic import DataFabricator
 from hotel_reservations.utility import create_parser, is_databricks, setup_logging
 
 # Setup environment and configurations
